adjust watering parameters based on practical tests
The water flow is really strong in my case and
the pots are relatively small.
Also, the water goes directly from the pipe end, there is no dripper,
sprinkler, etc.

So, I need to take longer pauses between watering "bursts" and
I need to stop watering earlier as it takes a while for the moisture to
get distributed around a pot.