(major) rework of the button handling and the display update
The button is now handled using OneButton library, a new dependency.
Previously the button was used to display current (RTC) time while the
button is pressed.  That was all.
Now we support 3 display modes:
- normal / default
- time
- technical.
A single button click now activates the new technical display mode
if the current mode is the default one.  For all other modes, a single
click returns to the default mode.

The technical mode currently displays raw sensor values in addition to
relative moisture levels (%%).  Also, it suppresses the pictograms.
This mode can be useful for sensor (re-)calibration.

The time display mode is activated via a double-click.
This works in any mode, but can be changed in the future.

The library also supports multi-click (3 or more) and long press.
Those are not assigned any functions now.

Other changes:
- more conservative variables types (byte instead of int,
  int instead of long);
- better names for some variables
- simplified logic for updating sensor values and the disply
- read_sensor() got inlined