re-calibrate sensors using dry and wet soil
One reading was taken in air-dried soil.
(Nothing extreme like the over-drying.)
The other was taken in water-saturated soil.
The water was added until it could no longer soak into the soil,
then excess water was removed.

Only one sensor was re-calibrated in this fashion, but the results are
used for all of them.